Update: naming of folders and structuring files properly
