+++ client/components/DeleteEvent.js
