+++ client/components/Gallery.js
