+++ client/components/PopularCenter.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
 
 /**
  * creates Navbar component
@@ -15,7 +16,7 @@ const PopularCenter = () => (
                         <div className="card-body">
                             <h4 className="card-title">Coming Of Kings</h4>
                             <p className="card-text">Some quick example text to build on the card title and make up the bulk of the card's content.</p>
-                            <a href="eventdetails.html" className="btn btn-mycolor btn-sm">info</a>
+                            <Link to="/eventdetails" className="btn btn-mycolor btn-sm">Details</Link>
                         </div>
                     </div>
                 </div>
@@ -25,7 +26,7 @@ const PopularCenter = () => (
                         <div className="card-body">
                             <h4 className="card-title">Coming Of Kings</h4>
                             <p className="card-text">Some quick example text to build on the card title and make up the bulk of the card's content.</p>
-                            <a href="eventdetails.html" className="btn btn-mycolor btn-sm">info</a>
+                            <Link to="/eventdetails" className="btn btn-mycolor btn-sm">Details</Link>
                         </div>
 
                     </div>
@@ -36,7 +37,7 @@ const PopularCenter = () => (
                         <div className="card-body">
                             <h4 className="card-title">Coming Of Kings</h4>
                             <p className="card-text">Some quick example text to build on the card title and make up the bulk of the card's content.</p>
-                            <a href="eventdetails.html" className="btn btn-mycolor btn-sm">info</a>
+                            <Link to="/eventdetails" className="btn btn-mycolor btn-sm">Details</Link>
                         </div>
 
                     </div>
