+++ client/components/app.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import LandingPage from '../components/LandingPage';
+
+/**
+ * Renders components to the page
+ * @returns {function} App
+ */
+const App = () => (
+    <div>
+        <LandingPage />
+    </div>
+);
+
+export default App;
