+++ client/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from 'react-dom';
-import { Router, browserHistory } from 'react-router';
 
-import routes from '../client/routes';
+import Root from './routes';
 
-render(<Router history={browserHistory} routes={routes}/>, document.getElementById('app'));
+
+render(<Root />, document.getElementById('app'));
