+++ package.json
@@ -47,10 +47,10 @@
         "mocha": "^4.0.1",
         "nodemon": "^1.12.1",
         "nyc": "^11.3.0",
-        "webpack": "^3.8.1",
         "webpack-dev-server": "^2.9.4"
     },
     "dependencies": {
+        "axios": "^0.17.1",
         "babel-cli": "^6.26.0",
         "babel-core": "^6.26.0",
         "babel-eslint": "^8.0.2",
@@ -63,12 +63,14 @@
         "bcrypt": "^1.0.3",
         "body-parser": "^1.18.2",
         "chai": "^4.1.2",
+        "webpack": "^3.8.1",
         "coveralls": "^3.0.0",
         "css-loader": "^0.28.7",
         "dotenv": "^4.0.0",
         "express": "^4.16.2",
         "extract-text-webpack-plugin": "^3.0.2",
         "file-loader": "^1.1.5",
+        "history": "^4.7.2",
         "image-webpack-loader": "^3.4.2",
         "jsonwebtoken": "^8.1.0",
         "morgan": "^1.9.0",
@@ -77,8 +79,11 @@
         "pg-hstore": "^2.3.2",
         "react": "^16.1.1",
         "react-dom": "^16.1.1",
-        "react-google-maps": "^9.4.1",
-        "react-router": "^3.2.0",
+        "react-redux": "^5.0.6",
+        "react-router-dom": "^4.2.2",
+        "react-router-redux": "^4.0.8",
+        "redux": "^3.7.2",
+        "redux-saga": "^0.16.0",
         "sass-loader": "^6.0.6",
         "sequelize": "^4.22.8",
         "sequelize-cli": "^3.1.0",
